Imported 'reusable' files from python nfl_picks
These files, while currently set up for the python version of nfl_picks,
will be the basis for this iteration of the web application. The HTML
files are all set up for jinja2 and will need to be converted to the
Go templating system. The scripts are also set-up for use in a
SQLAlchemy environment and will need to be converted as well.