Wrapped text/template to provide personalized navbars
In order to conditionally display the navbar depending on whether or not
a user was logged in, it was neccessary to supply the navbar template
with user information. To accomplish this, I wrapped text/template with
my own GoNflTemplate type and wrote an Execute for it that takes user
information and main frame content. This makes it much cleaner to add
this information whenever a template is executed.