Prevent nfl-app from crashing during the offseason
The nfl-pickem up was not functional during offesason times, period. A
lot of the logic was hardcoded to assume that we were in a current NFL
season. The is a complete rework of the application to allow the app to
be functional and testable even when the season is not current.

Key highlights:

- Remove the CLI application, most admin functionality is now in the
  webapp. Those features that are missing will be added.

- Return all errors in the database code, rather than panic.

- Rewrite all database code to properly use SQL to gather the exact
  information needed, rather than making multiple calls and manipulating
  the data manually.

- Introduce the concept of "off-season" to the app.

- Decouple logic in the handlers to allow for offseason and offline
  testing.

- Create testing code to generate a week of fake games based on the
  current date.